requirements.txt
Flask
Flask-SQLAlchemy
Pillow
python-dotenv